Add infrastructure for Prettier formatting compliance
A task and GitHub Actions workflow are added to provide compliance with the standardized code formatting style provided
by the Prettier formatter tool.

On every push and pull request that affects relevant files, the workflow runs the task to check whether the formatting
of supported files is compliant with the Prettier style.