Add infrastructure to lint YAML files
Add a task to lint the project's YAML files using the yamllint tool.

On every push and pull request that affects relevant files, a GitHub Actions workflow will run the task to check the
YAML files of the repository for issues.

The .yamllint.yml file is used to configure yamllint:
https://yamllint.readthedocs.io/en/stable/configuration.html