Add infrastructure to check for commonly misspelled words
Add tasks and a GitHub Actions workflow to check for commonly misspelled words.

On every push, pull request, and periodically, use codespell to check for commonly misspelled words.

In the event of a false positive, the problematic word should be added, in all lowercase, to the `ignore-words-list`
field of `.codespellrc`. Regardless of the case of the word in the false positive, it must be in all lowercase in the
ignore list. The ignore list is comma-separated with no spaces.