Add infrastructure for formatting TOML files
The project's Python package dependencies are configured via the pyproject.toml file. Since this file is edited
manually, it will be useful to provide infrastructure for standardizing the code formatting and validating compliance
with that standard.

This is done by the installation of the "prettier-plugin-toml", which adds support for TOML language to the Prettier
formatting tool. The `general:format-prettier` task will now also format the TOML files of the repository. The "Check
Prettier Formatting" workflow will enforce the formatting compliance.