Add infrastructure to test JavaScript/TypeScript code
A task and GitHub Actions workflow are added to use run the Jest-based tests for the project's JavaScript or TypeScript
code.

On every push and pull request that affects relevant files, and periodically, the workflow will run the repository's
TypeScript/JavaScript tests and upload the code coverage data to the Codecov service.

Codecov will comment a coverage report on all pull requests that trigger the workflow. This report will help the
contributor and reviewers to identify any introduced code that isn't being covered by tests.