Add npm package release infrastructure
In order to facilitate the use of the generator, the package is published to the npm registry.

A formal release procedure is defined in the project's Maintainer Guide.

A GiHub Actions workflow is added to automatically generate a GitHub release for every release tag, and publish the
package to the npm registry on every production release.