Use JSON schema to validate prompt data
The object exported by the prompt data file must have a specific data format. This data is provided by the user, so the
validity can not be assumed. It will be useful for the generator to do thorough validation of the data and provide
detailed communication to the user of any problems that are found.

Previously, some validation was done, but it was incomplete even with the current simple data format. That approach will
not scale as the expansion of the generator capabilities results in increased complexity of the prompt data format. The
standard way to accomplish this is by validating the data against a schema.

The schema is stored in a dedicated standard JSON schema file. In addition to the schema's use by the generator itself,
this also offers users the option to directly validate their prompt data files against the JSON schema, which might be
useful for their development and continuous integration systems.