Make "prompt data" plural where it refers to multiple prompts
The user configures prompts that are presented by the generator via a JavaScript file.

Previously this file was referred to as a "prompt data file", and its data as "prompt data". Although not incorrect,
this terminology made it difficult to differentiate between the array of configuration data for all the prompts and the
element for the configuration of an individual prompt within that array.

One approach would be to use "prompt datum" for the individual prompt data. However, this is somewhat inaccurate because
the individual configuration element still contains multiple items of data. In addition, although "datum" is used
strictly in other fields, "data" is used in both plural and singular contexts in computer science (and "datum" not at
all).

A more clear approach is to make "prompt" plural when referring to the configuration data for all the prompts: "prompts
data".