Support generation of complete front matter document from prompts
The generator is intended to be used with knowledge bases that are organized by tags. The tags are associated with a
document via the standard `tags` key of the Markdown front matter. Since the tags should be standardized for the KB, and
will often form related sets or alternatives, it will be convenient for the user to select the tags for a new document
via prompts.

Although it is possible to then reference the answers to these prompts directly in the document template to form the
front matter, this is prone to human error and also can require a significant amount of template content (especially
with the more advanced tagging use cases that will be supported by the generator in the future). In order to relieve the
user from the burden of developing and maintaining this template content in each dependent project, it will be helpful
for the generator to generate the full front matter content, making it available for insertion into the document by a
single simple variable reference in the template.