Disable internal anchor checks by link checker tool
The markdown-link-check tool is used to check for broken links in this project's Markdown files.

Validation of links to internal anchors was recently added to the tool.

Although a much needed feature, unfortunately it does not currently support anchors created by HTML anchor tags.

For example, this is valid and common Markdown:

[click here](#some-anchor)
<a name="some-anchor"></a>

but will fail the check:

ERROR: 1 dead links found!
[x] #some-anchor -> Status: 404

This type of link markup is used in cases where the anchors automatically generated for the headings might be ambiguous.

The solution is to configure markdown-link-check to skip these links.

This change should be be reverted whenever the tool is able to correctly validate all internal anchor links.