Add `mem-fs-editor` package dependency to "Acknowledgements" document
This is a transitive dependency via yeoman-generator, for which it provides important internal functionality. That alone
would not justify its inclusion (since, although they are important, the document would be huge and essentially rendered
meaningless if every transitive dependency was listed). However, `mem-fs-editor` is also exposed for nearly direct use
in the generator code via the `this.fs` wrapper, and it is used there for the essential purpose of template expansion.
For this reason, it merits mention in the acknowledgements.