Document appropriate EJS tag for use in templates
The snippets in the EJS documentation exclusively use the `<%= reference %>` tag format. This would lead the reader to
use the same tag format when writing their KB document templates. That tag format escapes the value as appropriate for
use in HTML code, which is not appropriate for use in the Markdown files produced by this generator. This will cause a
bad experience for the user at some point in the future when some specific input results in invalid generated document
content and they would then have some difficulty determining the cause, even after studying the minimal EJS
documentation. So it will be important to communicate this in the generator's template documentation.