Provide link to Inquirer `Question` object documentation
The configures arbitrary prompts for the generator in the "prompts data file" using the Question object of the Inquirer
package. It is essential for the user to understand the format of this object.

Previously, the generator's documentation provided a link to the documentation of the styles of the various prompt types
provided by Inquirer. That link does provide important information, but not the essential documentation of the actual
`Question` object format.

This documentation deficiency is resolved by adding a link to the documentation for the `Question` object, and also
making it clear that it is this specific object that is used in the `inquirer` property of the prompt data objects.