Wrap "..." placeholders in snippets with angle brackets
The code snippets in the readme use `...` placeholders to indicate additional content that is not shown. The convention
of wrapping explicit placeholders in angle brackets has been established, but previously was not used for these general
placeholders. Using the same format for all placeholders makes it a bit easier for the reader to differentiate them from
the actual code content of the snippets.