Expand documentation of kbDocumentTitle
The answer to the "built-in" "Knowledge base document title" prompt is available for use in the template by referencing
`kbDocumentTitle`.

Previously this was documented under the "built-in prompts" section of the readme. This organization would make sense if
there were multiple such prompts, but at this time there is only one so it doesn't.

The heading is hereby changed to be specific to this information. The previously very minimal documentation is expanded
to provide an explanation of the reason for it being built in as well as more clearly documenting how to use it in the
template and the recommended usage.