Document the usage of the array element append JSON pointer syntax in `frontMatterPath`
The primary use case for the generator's front matter document generation feature is assigning tags to documents based
on the prompt answers. The standard is to assign tags as an array of strings under the `tags` front matter key. So it
will be essential for the user to understand how to target an array using the `frontMatterPath` prompt data property.

The JSON pointer syntax provides the ability to append an element to an array. This syntax is not intuitive and the
documentation of the syntax in the specification is not easy to find. So it will be useful for the subject to be
explained in the generator's documentation.