Sort front matter data
Previously, the front matter data was ordered according to the prompts data file. Although the order is not likely of
importance to a machine consumer of the front matter data, a human will find it easier to read and maintain the front
matter if it is sorted in lexicographical order.

In case some users might prefer the sorting left according to the prompts data file, an option is added to the generator
configuration file for disabling this feature.