Expand "Test JavaScript" workflow paths filter to cover all relevant files
For the sake of efficiency the project's GitHub Actions workflows are configured to only trigger when relevant files are
modified.

Previously some paths of relevant files were not included in the paths filters of this workflow. That could have led to
important validation not being performed on contributions that modified those files.