Rename readme heading from "Usage" to "Generator Usage"
The initial motivation for this was as a workaround for a bug in a table of contents generator tool that was linking to
the "Configuration > Prompts Data File > `usage`" heading instead of the "Usage" heading.

However, I actually think the change has a beneficial effect on the readability since "Usage" might be a bit ambiguous
in the context of this readme that touches on multiple subjects.