Add infrastructure for generating and maintaining documentation table of contents
A task is added to generate a table of contents from the headings in the Markdown file. In addition to the initial
addition of the table of contents, this task can also be used to update the ToC after changes to the headings.

A GitHub Actions workflow is added to ensure the table of contents is kept in sync with the content. On every push or
pull request that affects the repository's Markdown files that contain a generated table of contents, and periodically
to catch external breakage, the workflow runs the task and then checks for a diff (which indicates an out of sync
condition).