Add answer processor feature
In some cases, it is necessary to make some transformations on the prompt answer value in order to make it suitable for
use in the document. This can be done in the template code, but this means the often complex EJS code must be written
and maintained in each knowledge base project. It will be more convenient for each of the commonly needed processing
operations to be implemented in the generator code instead.