Trim whitespace from answer values
When the generator user is entering free text at the prompts in the terminal, it is fairly easy to accidentally
introduce leading or trailing whitespace. This will then cause poor formatting in the created KB document. So it will be
helpful for the generator to trim the strings.

Although this could be implemented as a user-controlled processor, I can't imagine any use case where the user would
actually want to preserve leading or trailing whitespace in answers so I chose to just hardcode the behavior into the
generator.

The ability to programmatically add content (which can include whitespace) to answers will be provided via a "template"
answer processor.