Add dedicated documentation section re: "answer array"
The term "answer array" is used in multiple parts of the documentation. Previously it was explained in some detail in
two locations (resulting in some duplicate content), and only mentioned in passing without offering any reference to the
reader in the other locations.

This is a complex subject, and one that will be essential for any user of the "checkbox" prompt type or the "csv"
processor to understand. For this reason, a dedicated section is added to the documentation for the explanation of the
subject. The mentions of the subject in other parts of the documentation can now link to that information to provide the
reader with easy access to the relevant information.