Move information re: prompts data JSON schema to dedicated docs section
The project contains a JSON schema to validate the data format of the user's "prompts data" configuration. The primary
purpose of this component is to allow the generator itself to validate the data. However, it is possible that some users
might find it useful to be able to directly validate their prompts data files. For this reason, the JSON schema is
published in a dedicated file, and the availability of that file mentioned in the readme.

Previously this information was intentionally not given a dedicated sub-heading, instead simply placed at the bottom of
the "Prompts Data File" section of the readme under a horizontal rule. The reason for this is that it is likely that
very few users will be interested in this information, so the goal was to include it in the least intrusive manner
possible (ideally it would be placed in a separate "advanced documentation" document, but there isn't any other content
appropriate for such a document so it isn't worth creating one and adding a link to it at this time).

New content was later accidentally added below the location of the content about the JSON schema, causing it to be
placed inside an irrelevant sub-section. This error shows that the previous approach of not giving the distinct content
dedicated section makes the documentation harder for the maintainer and reader both to work with.