Apply "template" processor to elements of answer arrays
Some prompt configurations result in an array of answers.

Previously, the "template" processor always applied the template directly to the answer. That approach caused it to not
be of use with answer arrays.

The processor is made useful for answer arrays by applying the template to each of the elements of the array.