Add missing answers to tests
Evidently the generator test helpers allows the generator to run even when an answer is not set for one of the prompts.
Since these tests are covering exceptions, the lack of an answer was not impactful. However, it is a bad idea to rely on
the generator being able to run normally with missing data, and even when it is working the incomplete test
configuration might still cause contributors confusion. So it is best to always provide answers for all configured
prompts.