Remove duplicate definition of primary KB document filename in code
The filename is referenced multiple times in the code. In order to avoid duplication, a variable was defined that was
intended to be used exclusively in all code that needs this information. However, one of the existing references was not
updated at that time and so the filename was defined redundantly.