Split tests into two primary suites
Previously, all tests were under a single "running the generator" suite. This would make sense if there were to be some
unit tests that did not involve running the generator (e.g., testing functions directly). However, it seems that
JavaScript doesn't provide any good way of testing code without making it part of a public API. This project is only
intended to be an application, not a library. If I am going to expose a public API, I would want to invest resources
into the design and maintenance of that API within the context of it being used as a library rather than only
internally, and those resources are not available. So I am stuck with only doing integration tests. Even if I did
eventually add unit tests, the name of the current test file `integration.js` already defines it as a suite dedicated to
integration tests so I would simply create separate files for unit test suites.