Add document supplement creation feature
When the generator creates a new document, it contains a single "document primary file".

In some cases it might be useful to break a document into multiple files rather than having all the content in a single
file. Support for this use case is hereby provided by adding a built-in prompt that allows the user to select whether
they want to create a new document, or add a supplement file to an existing document.

The use of the generator to create document supplement files has similar benefits to those of using it to create the
document:

* Automatically creating a standardized filename based on the title
* Populating the file with a skeleton in the standardized format
* Populating the file with metadata and content based on answers to prompts