Don't modify user's generator configuration file
Universal generator user configuration is done via the .yo-rc.json file. Some of the configuration options are optional,
with the generator defaulting to the configuration that will be wanted by the majority of users. The Yeoman framework
offers what seems to be a convenient feature for implementing default configuration values. Unfortunately it turns out
the feature is not suitable for my use case. Instead of just setting the defaults values in the ephemeral data used by
the generator application during the run as I need, it writes the defaults to the user's configuration file. Perhaps the
feature would be useful for an application where it is used to generate a starting configuration for the user. However,
if the user already has a configuration it is incredibly hostile (and not done by any other application ever made) for
an application to unilaterally write to that file. This is also harmful because it means that there is no clean way for
the default values to be changed in later releases of the generator once the user has used a version with the previous
defaults.

It also has a significantly harmful effect on the UX because it causes a cryptic and alarming message to be printed to
the terminal on every generator run:

```
force .yo-rc.json
```

Since the design of Yeoman's default configuration feature is utterly idiotic, I must implement one myself.