Generalize and centralize prompt data default values implementation
Previously, the code for definining and handling default configuration values for user prompt data was sprinkled through
the code. This made it difficult to maintain.

The code was also implemented in a manner specific to each individual property. This made it difficult to expand.

It is now made into a general purpose default system. Default values are defined and their handling implemented in a
single location.