Assign prompts data to class property as early as possible in generator code
Previously, the prompts data was only stored in the dedicated class property at the end of the `initializing` function.
The code before the assignment accessed the data via a local variable, while all the code after accessed it via the
dedicated class property.

This made the code difficult to follow and bugs easy to introduce. That is avoided by assigning the property as early as
possible, then using it exclusively in all the rest of the code.