Set prompts data defaults before performing additional validations
The generator performs various validations on the user's prompt data file. Previously these were all performed before
applying the defaults to the data. It is best to perform the basic data structure validations on the user's data alone,
but when it comes to the additional more complex validations, where the requirements on one property are dependency on
another property, the correct validation result might only be obtained after the defaults are applied.

For this reason, the additional validation code is moved to after the code that sets the defaults.