Switch to "prompts configuration" term
The generator user can configure arbitrary prompts via a JavaScript file. Previously, this was referred to as the
"prompts data file", and the content as "prompts data" (when referring to the array of prompts) or "prompt data" (when
referring to a single prompt).

The use of "data" in the terms is potentially ambiguous since the intuitive interpretation is that this is referring to
the data from the answers the user makes to the prompts, rather than the data for configuring the prompts that will be
presented by the generator. The term "prompts configuration" is more clear and also aligns with the "generator
configuration file" and "generator configuration" terms used to refer to `.yo-rc.json` file and the configuration for
the `@per1234/generator-kb-document` generator provided by the file.