Use JSON schema to validate generator configuration
The user provides the universal generator configuration via the `@per1234/generator-kb-document` object in the
`.yo-rc.json` file. This object must have a specific data format. This data is provided by the user, so the validity can
not be assumed. It will be useful for the generator to do thorough validation of the data and provide detailed
communication to the user of any problems that are found.

There was some existing validation of the generator configuration, but not of its data format. When the configuration
had an incorrect format, it resulted in error messages that did not clearly indicate the real cause of the problem to
the user.

The validation is done in the standard way by validating the data against a schema. This is consistent with the
previously established approach used for validating the prompts configuration data format.

The schema is stored in a dedicated standard JSON schema file. In addition to the schema's use by the generator itself,
this also offers users the option to directly validate their generator configuration against the JSON schema, which
might be useful for their development and continuous integration systems.