Support providing answers via command line flags
For automated use cases (e.g., users making integration tests for their generator configuration), the generator's
interactive prompt interface is not appropriate.

Unfortunately Yeoman doesn't provide a built-in feature for setting answers via the generator invocation. This meant
that, previously, it was impossible to do a run of the generator without using an interactive interface.

However, `yo` does accept arbitrary command line flags, the data from which is made available for use in the generator
code. So I was able to implement the functionality (which really should be provided by the Yeoman codebase) in the
generator codebase. This was done by setting the answer for any prompt for which an answer had already been provided via
a flag in the `yo` invocation, and configuring that prompt to not be presented to the user. This means that if the user
provides flags for all prompts, the generator will run without requiring interaction.