Use prompt-appropriate type for answers from flags
The "checkbox" prompt type allows the user to select multiple choices. For this reason, this prompt type produces an
array of answers, rather than a single answer value as the other prompt types do.

Yeoman's command line flag parser produces an array of values when the same flag is passed multiple times in the `yo`
invocation. This allows the user to produce answer data of an appropriate type for a "checkbox" prompt via flags when
they wish to select multiple choices for the prompt. However, the flag system doesn't know anything about what the flag
data will be used for so it will not produce answer data of array type if the user only selects a single choice. The
flag system doesn't offer any way for the user to produce an array under these conditions.

So the type of answers from command line flags must be converted as appropriate for the prompt type by the generator.