Specify license type in package metadata
I had hoped that, in the absence of a `license` metadata field, npm would be smart enough to get the license type from
the canonical source that is the license file. That would have allowed me to avoid the duplication of this information.
Unfortunately, but I see that the npm package registry page for the package shows "license: none", and the `npm view
"@per1234/generator-kb-document"` output as "Proprietary".